// miniprogram/api/generated/models/AccountSecurityResponse.ts
// @ts-nocheck
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface AccountSecurityResponse
 */
export interface AccountSecurityResponse {
    /**
     * 
     * @type {string}
     * @memberof AccountSecurityResponse
     */
    phoneMasked?: string;
    /**
     * 
     * @type {boolean}
     * @memberof AccountSecurityResponse
     */
    phoneVerified?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof AccountSecurityResponse
     */
    passwordSet?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof AccountSecurityResponse
     */
    wechatBound?: boolean;
    /**
     * 
     * @type {string}
     * @memberof AccountSecurityResponse
     */
    accountName?: string;
    /**
     * 
     * @type {boolean}
     * @memberof AccountSecurityResponse
     */
    accountNameSet?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof AccountSecurityResponse
     */
    canChangeAccountName?: boolean;
    /**
     * 
     * @type {string}
     * @memberof AccountSecurityResponse
     */
    accountNameNextChangeAt?: string;
}

/**
 * Check if a given object implements the AccountSecurityResponse interface.
 */
export function instanceOfAccountSecurityResponse(value: object): value is AccountSecurityResponse {
    return true;
}

export function AccountSecurityResponseFromJSON(json: any): AccountSecurityResponse {
    return AccountSecurityResponseFromJSONTyped(json, false);
}

export function AccountSecurityResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): AccountSecurityResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'phoneMasked': json['phone_masked'] == null ? undefined : json['phone_masked'],
        'phoneVerified': json['phone_verified'] == null ? undefined : json['phone_verified'],
        'passwordSet': json['password_set'] == null ? undefined : json['password_set'],
        'wechatBound': json['wechat_bound'] == null ? undefined : json['wechat_bound'],
        'accountName': json['account_name'] == null ? undefined : json['account_name'],
        'accountNameSet': json['account_name_set'] == null ? undefined : json['account_name_set'],
        'canChangeAccountName': json['can_change_account_name'] == null ? undefined : json['can_change_account_name'],
        'accountNameNextChangeAt': json['account_name_next_change_at'] == null ? undefined : json['account_name_next_change_at'],
    };
}

export function AccountSecurityResponseToJSON(json: any): AccountSecurityResponse {
    return AccountSecurityResponseToJSONTyped(json, false);
}

export function AccountSecurityResponseToJSONTyped(value?: AccountSecurityResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'phone_masked': value['phoneMasked'],
        'phone_verified': value['phoneVerified'],
        'password_set': value['passwordSet'],
        'wechat_bound': value['wechatBound'],
        'account_name': value['accountName'],
        'account_name_set': value['accountNameSet'],
        'can_change_account_name': value['canChangeAccountName'],
        'account_name_next_change_at': value['accountNameNextChangeAt'],
    };
}